backend/main.py
"""
What this file does: Main entry point for the Crypto Butler backend server
What it receives as input: Command line arguments and environment variables
What it returns as output: Running Flask server with WebSocket support
"""

from flask import Flask
from flask_cors import CORS
from flask_socketio import SocketIO
from dotenv import load_dotenv
import os

from api.routes import APIRoutes
from api.websocket import WebSocketHandler
from utils.logger import ButlerLogger

def create_app():
    """
    Create and configure Flask application
    
    Returns:
        Configured Flask app instance
    """
    # TODO: Initialize Flask app with configuration
    pass

def main():
    """
    Main entry point for the backend server
    """
    # TODO: Start the backend server
    pass

if __name__ == '__main__':
    main()
