backend/requirements.txt
anthropic
web3
apscheduler
flask
flask-cors
flask-socketio
python-dotenv
requests
eth-account
